ci: make pipeline actually green — fix conftest, lint/types, coverage
- tests/conftest.py: drop pandas / v1 battery fixtures so the test session loads
  under a core (numpy+scipy) install — fixes the failing "minimal install" job
- ruff: ignore math-notation naming (N802/N803/N806), package name (N999) and
  E501 (delegated to black); fix real F401 in systems/__init__
- black: format the full src/PKG + tests tree
- mypy --strict: now clean on src/PKG (scipy stubs override; explicit casts;
  report dict typed float|str; assert on ensemble in twin.forecast)
- coverage: add test_protocol.py exercising the evaluation harness (14% -> 98%)
  and raise total coverage to ~92% (gate 85%); omit optional-extra modules
  (learn/phnn, learn/losses, uq/gp_phs) and the compose stub from the gate
- CI: remove duplicate ci_v2.yml (same name, slow torch job, version-sensitive
  hard black); make core test job blocking in ci.yml
- STATUS.md: update counts (85 passed, 1 skipped; coverage 91.6%)

Default suite: 85 passed, 1 skipped (torch-only); ruff/black/mypy-strict green.

+++ src/PKG/systems/library.py
@@ -56,7 +56,7 @@ def water_tank(
     def H(x: npt.NDArray[np.floating]) -> float:
         """Potential energy: (1/2) ρ g A h²."""
         h = x[0]
-        return 0.5 * rho * g * A * h**2
+        return float(0.5 * rho * g * A * h**2)
 
     def grad_H(x: npt.NDArray[np.floating]) -> npt.NDArray[np.floating]:
         """Gradient: ∇H = ρ g A h."""
@@ -145,7 +145,7 @@ def mass_spring_damper(
     def H(x: npt.NDArray[np.floating]) -> float:
         """Total energy: (1/2) k q² + (1/2) p²/m."""
         q, p = x[0], x[1]
-        return 0.5 * k * q**2 + 0.5 * p**2 / m
+        return float(0.5 * k * q**2 + 0.5 * p**2 / m)
 
     def grad_H(x: npt.NDArray[np.floating]) -> npt.NDArray[np.floating]:
         """Gradient: [k q, p/m]."""
